+++ src/coreaudio_backend/device.rs
@@ -1047,7 +1047,7 @@ impl CaptureDevice for CoreaudioCaptureDevice {
                             else if let Some(resampl) = &mut resampler {
                                 debug!("Adjusting resampler rate to {speed}.");
                                 if async_src {
-                                    if resampl.resampler.set_resample_ratio_relative(speed, true).is_err() {
+                                    if resampl.set_resample_ratio_relative(speed, true).is_err() {
                                         debug!("Failed to set resampling speed to {speed}.");
                                     }
                                 }