+++ src/file_backend/device.rs
@@ -329,11 +329,7 @@ fn capture_loop(
                 rate_adjust = speed;
                 if let Some(resampl) = &mut resampler {
                     if params.async_src {
-                        if resampl
-                            .resampler
-                            .set_resample_ratio_relative(speed, true)
-                            .is_err()
-                        {
+                        if resampl.set_resample_ratio_relative(speed, true).is_err() {
                             debug!("Failed to set resampling speed to {speed}");
                         }
                     } else {