+++ src/pipewire_backend/device.rs
@@ -1048,7 +1048,7 @@ impl CaptureDevice for PipeWireCaptureDevice {
                                 debug!("Requested to adjust capture speed to {speed}");
                                 if let Some(resampl) = &mut resampler {
                                     if async_src {
-                                        if resampl.resampler.set_resample_ratio_relative(speed, true).is_err() {
+                                        if resampl.set_resample_ratio_relative(speed, true).is_err() {
                                             debug!("Failed to set resampling speed to {}", speed);
                                         }
                                     } else {