+++ src/pulse_backend/device.rs
@@ -328,7 +328,7 @@ impl CaptureDevice for PulseCaptureDevice {
                                     rate_adjust = speed;
                                     if let Some(resampl) = &mut resampler {
                                         if async_src {
-                                            if resampl.resampler.set_resample_ratio_relative(speed, true).is_err() {
+                                            if resampl.set_resample_ratio_relative(speed, true).is_err() {
                                                 debug!("Failed to set resampling speed to {}", speed);
                                             }
                                         }